Refactor files and enable reaching for all tables
