+++ src/typeorm/entities/juso.entity.ts
@@ -1,0 +1,42 @@
+import { Column, Entity, PrimaryColumn } from 'typeorm'
+import { TRoadnameTableName } from '../../lib/sido'
+
+export function getJusoEntityByTableName(tableName: TRoadnameTableName) {
+  @Entity({ name: tableName, synchronize: false })
+  class JusoEntity {
+    @PrimaryColumn({ type: 'varchar' })
+    manage_number!: string
+
+    @Column({ type: 'varchar' })
+    roadname_code!: string
+
+    @Column({ type: 'varchar' })
+    eupmyeondong_number!: string
+
+    @Column({ type: 'varchar' })
+    basement!: string | null
+
+    @Column({ type: 'smallint' })
+    building_primary_number!: number | null
+
+    @Column({ type: 'smallint' })
+    building_secondary_number!: number | null
+
+    @Column({ type: 'varchar' })
+    basic_area_number!: string | null
+
+    @Column({ type: 'varchar' })
+    change_reason_code!: string | null
+
+    @Column({ type: 'varchar' })
+    notice_date!: string | null
+
+    @Column({ type: 'varchar' })
+    previous_roadname_address!: string | null
+
+    @Column({ type: 'varchar' })
+    has_detail!: string | null
+  }
+
+  return JusoEntity
+}
